+++ projects/demo/src/app/footer/footer.component.ts
@@ -1,0 +1,11 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'app-footer',
+  templateUrl: './footer.component.html',
+  styleUrls: ['./footer.component.css'],
+})
+export class FooterComponent {
+  @Input() version!: string;
+  @Input() year!: string;
+}