+++ src/app/github-icon/github-icon.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-github-icon',
-  templateUrl: './github-icon.component.html',
-  styleUrls: ['./github-icon.component.css']
-})
-export class GithubIconComponent {
-}